requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
requests>=2.32.3
pydantic>=2.9.2
python-dateutil>=2.9.0.post0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
streamlit>=1.28.0
flask>=3.0.0
pandas>=2.0.0
